Add files via upload
add ncl script that creates individual panels for Figure 1 of Lesley Thorne's review paper on sea birds, "A review of the effects of wind on the movement, behavior, energetics, and life history of seabirds", published in Marine Ecology Progress Series.

This script plots wind values from ERA5 on the 10m height level.  Wind magnitude are shown with filled contours, continents and wind streamlines are shown in black, and lat/lon lines are shown in gray.